pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^5.0"
python = "^3.10"
pandas = "2.2.0"
lxml = "5.1.0"
django-import-export = "3.3.7"
[tool.poetry.dev-dependencies]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

# [tool.ruff]
# # https://beta.ruff.rs/docs/configuration/
# select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
# ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"